Update summary table from metrics tables
