+++ mlip_testing/app/utils/register_callbacks.py
@@ -1,0 +1,214 @@
+"""Register callbacks relating to interative weights."""
+
+from __future__ import annotations
+
+from dash import Input, Output, State, callback, ctx
+from dash.exceptions import PreventUpdate
+
+from mlip_testing.analysis.utils.utils import calc_ranks, calc_scores
+
+
+def register_table_callbacks(table_id) -> None:
+    """
+    Register callback to update table scores/rankings when stored values change.
+
+    Parameters
+    ----------
+    table_id
+        ID for table to update.
+    """
+
+    @callback(
+        Output(table_id, "data"),
+        Input(f"{table_id}-weight-store", "data"),
+        State(table_id, "data"),
+        prevent_initial_call=True,
+    )
+    def update_table_scores(
+        stored_weights: dict[str, float], table_data: list[dict]
+    ) -> list[dict]:
+        """
+        Update scores table score and rankings when data store updates.
+
+        Parameters
+        ----------
+        stored_weights
+            Stored weight values for `table_id`.
+        table_data
+            Data from `table_id` to be updated.
+
+        Returns
+        -------
+        list[dict]
+            Updated table data.
+        """
+        table_data = calc_scores(table_data, stored_weights)
+        return calc_ranks(table_data)
+
+    @callback(
+        Output("summary-table-scores-store", "data"),
+        Input(f"{table_id}-weight-store", "data"),
+        State(table_id, "data"),
+        State("summary-table-scores-store", "data"),
+        prevent_initial_call=False,
+    )
+    def update_scores_store(
+        stored_weights: dict[str, float],
+        table_data: list[dict],
+        scores_data: dict[str, list[float]],
+    ) -> dict[str, list[float]]:
+        """
+        Update stored scores values when weights update.
+
+        Parameters
+        ----------
+        stored_weights
+            Stored weight values for `table_id`.
+        table_data
+            Data from `table_id` to be updated.
+        scores_data
+            Dictionary of scores for each tab.
+
+        Returns
+        -------
+        dict[str, list[float]]
+            List of scoress indexed by table_id.
+        """
+        # Update scores store
+        return {table_id.removesuffix("-table"): [row["Score"] for row in table_data]}
+
+    @callback(
+        Output("summary-table", "data"),
+        Input("all-tabs", "value"),
+        Input("summary-table-weight-store", "data"),
+        State("summary-table-scores-store", "data"),
+        State("summary-table", "data"),
+        prevent_initial_call=False,
+    )
+    def update_summary_table(
+        tabs_value: str,
+        stored_weights: dict[str, float],
+        stored_scores: dict[str, list[float]],
+        summary_data: list[dict],
+    ) -> list[dict]:
+        """
+        Update summary table when scores change.
+
+        Parameters
+        ----------
+        tabs_value
+            Value of selected tab. Parameter unused, but required to register Input.
+        stored_weights
+            Stored summary weights dictionary.
+        stored_scores
+            Stored scores for table scores.
+        summary_data
+            Data from summary table to be updated.
+
+        Returns
+        -------
+        list[dict]
+            Updated summary table data.
+        """
+        # Update table from stored scores
+        if stored_scores:
+            for key, values in stored_scores.items():
+                for row, value in zip(summary_data, values, strict=True):
+                    row[key] = value
+
+        # Update table contents
+        summary_data = calc_scores(summary_data, stored_weights)
+        return calc_ranks(summary_data)
+
+
+def register_weight_callbacks(input_id: str, table_id: str, column: str) -> None:
+    """
+    Register all callbacks for weight inputs.
+
+    Parameters
+    ----------
+    input_id
+        ID prefix for slider and input box.
+    table_id
+        ID for table. Also used to identify reset button and weight store.
+    column
+        Column header corresponding to slider and input box.
+    """
+    default_weight = 1.0
+
+    @callback(
+        Output(f"{table_id}-weight-store", "data", allow_duplicate=True),
+        Input(f"{input_id}-slider", "value"),
+        Input(f"{input_id}-input", "value"),
+        Input(f"{table_id}-reset-button", "n_clicks"),
+        State(f"{table_id}-weight-store", "data"),
+        prevent_initial_call=True,
+    )
+    def store_slider_value(
+        slider_weight: float,
+        input_weight: float,
+        n_clicks: int,
+        stored_weights: dict[str, float],
+    ) -> dict[str, float]:
+        """
+        Store weight values from slider and text input.
+
+        Parameters
+        ----------
+        slider_weight
+            Weight value from slider.
+        input_weight
+            Weight value from input box.
+        n_clicks
+            Number of clicks. Variable unused, but Input is required to reset weights.
+        stored_weights
+            Stored weights dictionary.
+
+        Returns
+        -------
+        dict[str, float]
+            Stored weights for each slider.
+        """
+        trigger_id = ctx.triggered_id
+
+        if trigger_id == f"{input_id}-slider":
+            stored_weights[column] = slider_weight
+        elif trigger_id == f"{input_id}-input":
+            if input_weight:
+                stored_weights[column] = input_weight
+            else:
+                raise PreventUpdate
+        elif trigger_id == f"{table_id}-reset-button":
+            stored_weights.update((key, default_weight) for key in stored_weights)
+            stored_weights[column] = default_weight
+        else:
+            raise PreventUpdate
+
+        return stored_weights
+
+    @callback(
+        Output(f"{input_id}-input", "value"),
+        Output(f"{input_id}-slider", "value"),
+        Input(f"{table_id}-weight-store", "data"),
+        Input("all-tabs", "value"),
+        prevent_initial_call="initial_duplicate",
+    )
+    def sync_slider_inputs(
+        stored_weights: dict[str, float], tabs_value: str
+    ) -> tuple[float, float]:
+        """
+        Sync weight values between slider and text input via Store.
+
+        Parameters
+        ----------
+        stored_weights
+            Stored weight values for each column.
+        tabs_value
+            Tab name. Variable unused, but required as input to trigger on tab change.
+
+        Returns
+        -------
+        tuple[float, float]
+            Weights to set slider value and text input value.
+        """
+        return stored_weights[column], stored_weights[column]
